requirements.txt
Flask	2.1.2
Flask-SQLAlchemy	2.5.1
Flask-Session	0.4.0
Jinja2	3.1.2
MarkupSafe	2.1.1
SQLAlchemy	1.4.36
Werkzeug	2.1.2
cachelib	0.7.0
certifi	2022.5.18
charset-normalizer	2.0.12
click	8.1.3
colorama	0.4.4
greenlet	1.1.2
gunicorn	20.1.0
idna	3.3
itsdangerous	2.1.2
pip	21.1.2
python-dotenv	0.20.0
requests	2.27.1
setuptools	57.0.0
urllib3	1.26.9
wheel	0.36.2